requirements.txt
Pillow>=9.0.0
schedule>=1.2.0
pystray>=0.19.0
